Refactor functions into summary table, remove unused module filtering
